읽은 도서 저장 기능 ing

+++ src/main/java/com/project/backend/controller/MyLibraryController.java
@@ -14,6 +14,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/library")
@@ -25,9 +27,8 @@ public class MyLibraryController {
     private final LibraryService libraryService;
     private final LibraryRepository libraryRepository;
 
-    @PostMapping("/save-books")
+    @PostMapping("/save")
     public ResponseEntity<?> saveBooks(@RequestBody LibraryRequestDto request, HttpServletRequest httpServletRequest) {
-
         // JWT 토큰 검증
         String token = httpServletRequest.getHeader("Authorization").substring(7);
 
@@ -35,28 +36,75 @@ public class MyLibraryController {
             return new ResponseEntity<>("로그인 해주세요.", HttpStatus.UNAUTHORIZED);
         }
 
-        Long id = tokenProvider.getSubject(token);
+        Long id = tokenProvider.getSubject(token);  // 로그인한 회원의 longId 값
         User user = userRepository.findUserById(id);
-        String userId = user.getUserId();
 
         if (user == null) {
             return new ResponseEntity<>("유효하지 않은 사용자", HttpStatus.UNAUTHORIZED);
         }
 
-        libraryService.saveBook(id, userId, request);
+        libraryService.saveBook(id, request);
 
         return new ResponseEntity<>("내 서재에 저장되었습니다.", HttpStatus.OK);
     }
 
-    @GetMapping("/delete-book/{userId}/{isbn}")
-    public ResponseEntity<Void> deleteBook(@PathVariable Long num, @PathVariable String isbn) {
+    @GetMapping("/delete/{isbn}")
+    public ResponseEntity<?> deleteBook(@PathVariable String isbn, HttpServletRequest httpServletRequest) {
+        // JWT 토큰 검증
+        String token = httpServletRequest.getHeader("Authorization").substring(7);
+
+        if (token == null) {
+            return new ResponseEntity<>("로그인 해주세요.", HttpStatus.UNAUTHORIZED);
+        }
+
+        Long num = tokenProvider.getSubject(token);  // 로그인한 회원의 longId 값
+
         libraryService.deleteBook(num, isbn);
         return ResponseEntity.noContent().build();
     }
 
-    @GetMapping("/book/{Id}")
-    public ResponseEntity<List<Library>> getBooksById(@PathVariable Long num) {
+    @GetMapping("/books")
+    public ResponseEntity<?> getBooksById(HttpServletRequest httpServletRequest) {
+        // JWT 토큰 검증
+        String token = httpServletRequest.getHeader("Authorization").substring(7);
+
+        if (token == null) {
+            return new ResponseEntity<>("로그인 해주세요.", HttpStatus.UNAUTHORIZED);
+        }
+
+        Long id = tokenProvider.getSubject(token);  // 로그인한 회원의 longId 값
+        User user = userRepository.findUserById(id);
+
+        if (user == null) {
+            return new ResponseEntity<>("유효하지 않은 사용자", HttpStatus.UNAUTHORIZED);
+        }
+
+        Long num = user.getId();
+
         List<Library> books = libraryService.getBooksById(num);
         return ResponseEntity.ok(books);
+    }
+
+    @GetMapping("/check/{isbn}")
+    public ResponseEntity<?> checkBookExists(HttpServletRequest httpServletRequest, @PathVariable String isbn) {
+
+        // 토큰 검증
+        String token = httpServletRequest.getHeader("Authorization").substring(7);
+
+        if (token == null) {
+            return new ResponseEntity<>("로그인 해주세요.", HttpStatus.UNAUTHORIZED);
+        }
+
+        Long id = tokenProvider.getSubject(token);  // 로그인한 회원의 longId 값
+        User user = userRepository.findUserById(id);
+
+        if (user == null) {
+            return new ResponseEntity<>("유효하지 않은 사용자", HttpStatus.UNAUTHORIZED);
+        }
+
+        // LibraryRepository에서 도서 존재 여부 확인
+        Optional<Library> book = libraryService.checkBooks(user.getId(), isbn);
+
+        return ResponseEntity.ok(Map.of("exists", book.isPresent()));
     }
 }
